refactor :  js=>ts, jsx=>tsx 변환
