+++ .eslintrc.json
@@ -27,8 +27,9 @@
     "rules": {
         "eqeqeq": ["error", "always"],
         "global-require": 0,
-        // "@typescript-eslint/no-explicit-any": "error",
-        "@typescript-eslint/explicit-function-return-type":"off",
+        "consistent-return": "off",
+        "@typescript-eslint/no-explicit-any": "warn",
+        "@typescript-eslint/explicit-function-return-type":"error",
         "@typescript-eslint/adjacent-overload-signatures": "warn",
         "@typescript-eslint/no-dupe-class-members": "warn",
         "@typescript-eslint/no-empty-function": "warn",
