+++ src/api/user/types.ts
@@ -1,0 +1,3 @@
+export interface IDataTosubmitProps {
+  dataTosubmit: string;
+}
