+++ src/pages/Login.tsx
@@ -1,12 +1,12 @@
 import { Avatar, Box, Grid, Link, Typography, Container, CssBaseline } from "@mui/material";
-import naverImg from "assets/image/naver_oauth.png";
-import kakaoImg from "assets/image/kakao_oauth.png";
+import naverImg from "assets/img/naver_oauth.png";
+import kakaoImg from "assets/img/kakao_oauth.png";
 import * as Styled from "./style";
 import LoginForm from "components/LoginForm/LoginForm";
 import { kakaOauthUrl } from "utils/kakaoOAuthUtil";
 import { naverOauthUrl } from "utils/naverOAuthUtil";
 
-const LogIn = () => {
+const LogIn = (): JSX.Element => {
   return (
     <Container component="main" maxWidth="xs">
       <CssBaseline />
